Move SubmissionResult, Verdict, is_RTE, is_TLE to a new module, judge
