+++ problemtools/judge/__init__.py
@@ -1,0 +1,13 @@
+from .result import (
+    SubmissionResult,
+    Verdict,
+    is_RTE,
+    is_TLE,
+)
+
+__all__ = [
+    'SubmissionResult',
+    'Verdict',
+    'is_RTE',
+    'is_TLE',
+]
